Remove default preprocessing
Instead of providing preprocessing functionality by default, a callable function must now be passed at initialisation. This function is a user-defined text preprocessing method, and is called when parsing the documents.

Changed the dictionary keys so that they use a str instead of an int. This means that instead of using an ID to identify documents, a name will be used instead (e.g. the name of the webpage the document was retrieved from).